pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "speculative-whisper"
version = "0.1.0"
description = "Speculative decoding for OpenAI Whisper — Tiny drafts, Large V3 verifies."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Your Name", email = "you@example.com"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "openai-whisper>=20231117",
    "torch>=2.0",
    "torchaudio>=2.0",
    "numpy>=1.24",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "jiwer>=3.0",
    "fastapi>=0.100",
    "uvicorn[standard]>=0.22",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "httpx>=0.24",
    "mypy>=1.4",
    "ruff>=0.1",
]

[project.scripts]
speculative-whisper = "speculative_whisper.cli:main"

[tool.setuptools.packages.find]
include = ["speculative_whisper*", "api*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.ruff]
target-version = "py39"
line-length = 100
